Ver Front

+++ AppResta/AppResta/Model/Cart.cs
@@ -10,7 +10,7 @@ namespace AppResta.Model
         public string platillo { get; set; }
         public int cantidad { get; set; }
         public double precio { get; set; }
-        public int total { get; set; }
+        public double total { get; set; }
         public string estado { get; set; }
